+++ otelconf/x/generated_config.go
@@ -509,7 +509,7 @@ type ExperimentalComposableSampler struct {
 	//
 	RuleBased *ExperimentalComposableRuleBasedSampler `json:"rule_based,omitempty,omitzero" yaml:"rule_based,omitempty" mapstructure:"rule_based,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type ExperimentalContainerResourceDetector map[string]interface{}
@@ -997,7 +997,7 @@ type ExperimentalResourceDetector struct {
 	//
 	Service ExperimentalServiceResourceDetector `json:"service,omitempty,omitzero" yaml:"service,omitempty" mapstructure:"service,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type ExperimentalServiceResourceDetector map[string]interface{}
@@ -1224,7 +1224,7 @@ type LogRecordExporter struct {
 	//
 	OTLPHttp *OTLPHttpExporter `json:"otlp_http,omitempty,omitzero" yaml:"otlp_http,omitempty" mapstructure:"otlp_http,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type LogRecordLimits struct {
@@ -1266,7 +1266,7 @@ type LogRecordProcessor struct {
 	//
 	Simple *SimpleLogRecordProcessor `json:"simple,omitempty,omitzero" yaml:"simple,omitempty" mapstructure:"simple,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type LoggerProvider struct {
@@ -1325,7 +1325,7 @@ type MetricProducer struct {
 	//
 	Opencensus OpenCensusMetricProducer `json:"opencensus,omitempty,omitzero" yaml:"opencensus,omitempty" mapstructure:"opencensus,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type MetricReader struct {
@@ -1781,7 +1781,7 @@ type OpenTelemetryConfiguration struct {
 	//
 	TracerProvider *TracerProvider `json:"tracer_provider,omitempty,omitzero" yaml:"tracer_provider,omitempty" mapstructure:"tracer_provider,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 // Configure if the SDK is disabled or not.
@@ -1899,7 +1899,7 @@ type PullMetricExporter struct {
 	//
 	PrometheusDevelopment *ExperimentalPrometheusMetricExporter `json:"prometheus/development,omitempty,omitzero" yaml:"prometheus/development,omitempty" mapstructure:"prometheus/development,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type PullMetricReader struct {
@@ -1940,7 +1940,7 @@ type PushMetricExporter struct {
 	//
 	OTLPHttp *OTLPHttpMetricExporter `json:"otlp_http,omitempty,omitzero" yaml:"otlp_http,omitempty" mapstructure:"otlp_http,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type Resource struct {
@@ -2020,7 +2020,7 @@ type Sampler struct {
 	//
 	TraceIDRatioBased *TraceIDRatioBasedSampler `json:"trace_id_ratio_based,omitempty,omitzero" yaml:"trace_id_ratio_based,omitempty" mapstructure:"trace_id_ratio_based,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type SeverityNumber string
@@ -2085,7 +2085,7 @@ type SpanExporter struct {
 	//
 	OTLPHttp *OTLPHttpExporter `json:"otlp_http,omitempty,omitzero" yaml:"otlp_http,omitempty" mapstructure:"otlp_http,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type SpanKind string
@@ -2179,7 +2179,7 @@ type SpanProcessor struct {
 	//
 	Simple *SimpleSpanProcessor `json:"simple,omitempty,omitzero" yaml:"simple,omitempty" mapstructure:"simple,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type SumAggregation map[string]interface{}
@@ -2215,7 +2215,7 @@ type TextMapPropagator struct {
 	//
 	Tracecontext TraceContextPropagator `json:"tracecontext,omitempty,omitzero" yaml:"tracecontext,omitempty" mapstructure:"tracecontext,omitempty"`
 
-	AdditionalProperties interface{} `mapstructure:",remain"`
+	AdditionalProperties map[string]any `mapstructure:",remain" json:"-" yaml:"-"`
 }
 
 type TraceContextPropagator map[string]interface{}